powinno ale sie nie uczy

+++ traffic fluid simulator/backend/env_4 — kopia/model/Net.py
@@ -12,7 +12,7 @@ class Times:
 class Net: # to jest dawane na front
     times: Times
     rewards: List[float] # 3 liczby
-    actions: List[float] # 3 liczby
+    actions: List[int] # 3 liczby
     densities: List[float] # 36 liczb
     lights: List[List[float]] = attr.ib(factory=list) # 36 x 36 liczb
